PHOENIX-5456 IndexScrutinyTool slow for indexes on multitenant tables and IndexScrutinyIT doesn't run
